router.hpp and router.cpp
