+++ refactor/include/Http_Response.hpp
@@ -1,0 +1,3 @@
+#pragma once
+
+struct Response {};
